implementing exception handler

+++ src/main/java/com/codesgm/sales/services/UserService.java
@@ -2,7 +2,10 @@ package com.codesgm.sales.services;
 
 import com.codesgm.sales.entities.User;
 import com.codesgm.sales.respositories.UserRepository;
+import com.codesgm.sales.services.exceptions.DatabaseException;
+import com.codesgm.sales.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +23,7 @@ public class UserService {
 
     public User findById(Long id){
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public  User insert(User user) {
@@ -28,7 +31,17 @@ public class UserService {
     }
 
     public void delete(Long id){
-         repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        } catch (RuntimeException e) {
+            e.printStackTrace();
+        }
     }
 
     public User update(Long id, User user) {
